Move TimeQueue into PlaybackSchedule and improve the comments
