+++ src/AudioIO.h
@@ -448,29 +448,6 @@ public:
 
 protected:
    RecordingSchedule mRecordingSchedule{};
-
-   // Another circular buffer
-   // Holds track time values corresponding to every nth sample in the playback
-   // buffers, for some large n
-   struct TimeQueue {
-      Doubles mData;
-      size_t mSize{ 0 };
-      double mLastTime {};
-      // These need not be updated atomically, because we rely on the atomics
-      // in the playback ring buffers to supply the synchronization.  Still,
-      // align them to avoid false sharing.
-      struct Cursor {
-         size_t mIndex {};
-         size_t mRemainder {};
-      };
-      NonInterfering<Cursor> mHead, mTail;
-
-      void Producer(
-         const PlaybackSchedule &schedule, double rate, double scrubSpeed,
-         size_t nSamples );
-      double Consumer( size_t nSamples, double rate );
-   } mTimeQueue;
-
    PlaybackSchedule mPlaybackSchedule;
 
 private:
